Added support for removing the Qdisc filter on load
With this commit, the Qdisc filter will be removed from the listening
device when the program loads. However, if the user provides the '-q'
flag, the program will fail to load if a Qdisc filter is on the device's
ingress Qdisc.

$ neighsnoopd -q -v br_default
libbpf: Kernel error message: Exclusivity flag on, cannot modify
libbpf: Kernel error message: Filter already exists

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>