Added support for device linking lookup
This commit adds a lookup where the IP address is matched against local
network interfaces to see if it is on the same network as that device.
If it is, it will now check the IFLA_LINK field of the device to confirm
that it has a link to the monitored bridge device.

It also adds extra Netlink logging for RTM_NEWLINK and RTM_GETLINK
packets.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>